Update `devDependencies`, break up and organize code
