+++ src/index.ts
@@ -1,0 +1,16 @@
+export type { IPrintContextProps } from "./components/PrintContext";
+export { PrintContextConsumer } from "./components/PrintContext";
+export { ReactToPrint } from "./components/ReactToPrint";
+export { useReactToPrint } from "./hooks/useReactToPrint";
+export type { IReactToPrintProps, ITriggerProps } from "./types/reactToPrintProps";
+
+import { ReactToPrint } from "./components/ReactToPrint";
+export default ReactToPrint;
+
+
+
+
+
+
+
+
