+++ src/types/font.ts
@@ -1,0 +1,7 @@
+// https://developer.mozilla.org/en-US/docs/Web/API/FontFace/FontFace
+export type Font = {
+    family: string;
+    source: string;
+    weight?: string;
+    style?: string;
+};
