+++ webpack.prod.config.js
@@ -6,7 +6,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
 module.exports = {
     mode: 'production',
-    entry:  './src/index.tsx',
+    entry:  './src/index.ts',
     output: {
         environment: {
             // The environment supports arrow functions ('() => { ... }').
